requirements.txt
google-generativeai==0.8.3
langchain==0.3.3
PyPDF2==3.0.1
faiss-cpu==1.9.0
langchain_google_genai==2.0.1
langchain-community==0.3.2
Flask==3.0.3
fpdf==1.7.2
librosa==0.10.2.post1
matplotlib
numpy
openai==1.47.1
pandas
praat-parselmouth==0.4.4
scikit-learn
scipy
SpeechRecognition==3.10.4
flask_cors==5.0.0
supabase==2.7.4
python-dotenv==1.0.1
pydub==0.25.1
gunicorn==23.0.0
langdetect==1.0.9